move utils functions to
